requirements.txt
torch>=2.0
torchvision
ultralytics
pillow
tensorflow
streamlit

